requirements.txt
PyQt6==6.4.0
python-vlc==3.0.18122
opencv-python==4.8.1.78
python-dotenv==1.0.0
